refactor: 트랜잭션 적용 통일성을 위한 메서드 구조 수정
