+++ src/main/java/wad/seoul_nolgoat/auth/oauth2/security/CustomOAuth2UserService.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wad.seoul_nolgoat.auth.oauth2.dto.GoogleResponse;
 import wad.seoul_nolgoat.auth.oauth2.dto.KakaoResponse;
+import wad.seoul_nolgoat.auth.service.AuthService;
 import wad.seoul_nolgoat.exception.ApiException;
-import wad.seoul_nolgoat.service.user.UserService;
 
 import static wad.seoul_nolgoat.exception.ErrorCode.UNSUPPORTED_PROVIDER;
 
@@ -24,7 +24,7 @@ public class CustomOAuth2UserService extends DefaultOAuth2UserService {
     public static final String GOOGLE = "google";
     public static final String PROVIDER_ID_DELIMITER = "_";
 
-    private final UserService userService;
+    private final AuthService authService;
 
     @Transactional
     @Override
@@ -33,10 +33,10 @@ public class CustomOAuth2UserService extends DefaultOAuth2UserService {
 
         String registrationId = userRequest.getClientRegistration().getRegistrationId();
         if (registrationId.equals(KAKAO)) {
-            return userService.processOAuth2User(new KakaoResponse(oAuth2User.getAttributes()));
+            return authService.processOAuth2User(new KakaoResponse(oAuth2User.getAttributes()));
         }
         if (registrationId.equals(GOOGLE)) {
-            return userService.processOAuth2User(new GoogleResponse(oAuth2User.getAttributes()));
+            return authService.processOAuth2User(new GoogleResponse(oAuth2User.getAttributes()));
         }
         log.error("Unsupported OAuth2 provider: {}", registrationId);
         throw new ApiException(UNSUPPORTED_PROVIDER);
